empty gallery work

+++ public/includes/components/component-gallery.php
@@ -44,11 +44,12 @@ class AesopCoreGallery {
 		//gallery caption
 		$gallery_caption = get_post_meta( $gallery_id, 'aesop_gallery_caption', true);
 
+
 		ob_start();
 
 			do_action('aesop_gallery_before', $type, $gallery_id); //action
 
-			?><div id="aesop-gallery-<?php echo esc_attr( $unique );?>" <?php echo aesop_component_data_atts( 'gallery', $gallery_id, $atts );?> class="aesop-component aesop-gallery-component aesop-<?php echo esc_attr($type);?>-gallery-wrap"><?php
+			?><div id="aesop-gallery-<?php echo esc_attr( $unique );?>" <?php echo aesop_component_data_atts( 'gallery', $gallery_id, $atts );?> class="aesop-component aesop-gallery-component aesop-<?php echo esc_attr($type);?>-gallery-wrap <?php if ( empty( $gallery_id ) ) echo 'empty-gallery';?> "><?php
 
 				do_action('aesop_gallery_inside_top', $type, $gallery_id); //action
 
@@ -79,17 +80,30 @@ class AesopCoreGallery {
 						printf('<p class="aesop-component-caption">%s</p>', esc_html( $gallery_caption ) );
 					}
 
-					if ( is_user_logged_in() ) {
+					// provide the edit link to the backend edit if Aesop Editor is not active
+					if ( !class_exists('Aesop_Editor') && is_user_logged_in() && current_user_can('edit_posts') ) {
 						$url = admin_url( 'post.php?post='.$gallery_id.'&action=edit' );
 						$edit_gallery = __('edit gallery', 'aesop-core');
 						printf('<a class="aesop-gallery-edit aesop-content" href="%s" target="_blank" title="%s">(%s)</a>',$url, $edit_gallery, $edit_gallery );
 					}
 
-				} elseif( empty( $image_ids ) || empty( $gallery_id ) ) {
+				}
 
-					?><div class="aesop-error aesop-content"><?php
-						_e('This gallery is empty! It\'s also possible that you simply have the wrong gallery ID.', 'aesop-core');
-					?></div><?php
+				if ( empty( $gallery_id ) && is_user_logged_in() && current_user_can('edit_posts') ) {
+
+					if ( class_exists('Aesop_Editor') ) {
+
+						?><div class="aesop-editor--empty-gallery"><?php
+							_e('Setup a gallery by clicking the <span class="aesop-icon-gear"></span> icon above.', 'aesop-core');
+						?></div><?php
+
+					} else {
+
+						?><div class="aesop-error aesop-content"><?php
+							_e('This gallery is empty! It\'s also possible that you simply have the wrong gallery ID.', 'aesop-core');
+						?></div><?php
+
+					}
 				}
 
 				do_action('aesop_gallery_inside_bottom', $type, $gallery_id); //action
